Force NKRO.
I do not care about the keyboard not working in the BIOS, or during
early boot. But I do not have bootmagic enabled, and the magic combo to
turn on NKRO is not convenient, because my layout doesn't have a right
shift to begin with.

For these reasons, force NKRO on by default. Now I can use long chords
in Plover!

Signed-off-by: Gergely Nagy <algernon@madhouse-project.org>