UCIS: Beginnings of an unicode input system
Work in progress, towards #84. What this does, is you enable the UCIS
mode, and then it will record the next few keys, until one hits Enter or
Space. At that point, it will try to look up the symbol from a lookup
table, and do some unicode input magic to make it happen. If the lookup
fails, it will just replay the keys.

This is a work in progress, with a number of things that can be
optimized and/or cleaned up still.

Signed-off-by: Gergely Nagy <algernon@madhouse-project.org>