Guard the function key timers a bit more
When trying to decide whether to fire a number or a function key, make
sure that the timer is active, before firing a function key. For
example, if the key was pressed in another layer, held a bit and
released in the base layer, then previously we would have fired a
function key. Now we fire the number key instead.

This fixes #87.

Signed-off-by: Gergely Nagy <algernon@madhouse-project.org>